requirements.txt
django==2.2.3
confluent-kafka==0.11.6
django-redis==4.10.0
python-bitcoinrpc===1.0
python-memcached
requests===2.22.0
cassandra-driver===3.18.0
 django-cors-headers===3.1.0